Add extraction of population MCMC extimates from SS output * Add crayon and clisymbols to package

+++ R/ss-model-loading.R
@@ -1,39 +1,61 @@
 #' Create an rds file to hold an SS model's data and outputs.
 #'
 #' @param model_dir Directory name of model to be loaded
+#' @param overwrite_ss_rds Logical. Overwrite the RDS file if it exists
 #' @param ... Absorb arguments destined for other functions
 #'
 #' @return [base::invisible()]
 #' @importFrom r4ss SSgetMCMC
 #' @export
 create_rds_file <- function(model_dir = NULL,
+                            overwrite_ss_rds = TRUE,
                             ...){
 
   verify_argument(model_dir, "character", 1)
+  verify_argument(overwrite_ss_rds, "logical", 1)
 
   if(!dir.exists(model_dir)){
     stop("Error - the directory ", model_dir, " does not exist.\n",
          "Fix the problem and try again.", call. = FALSE)
   }
 
-  # The RDS file will have the same name as the directory it is in
   rds_file <- file.path(model_dir, paste0(basename(model_dir), ".rds"))
-  if(file.exists(rds_file)){
+  # The RDS file will have the same name as the directory it is in
+  if(file.exists(rds_file) && overwrite_ss_rds){
     unlink(rds_file, force = TRUE)
   }
-
-  message("Creating a new RDS file in ", model_dir, "\n")
-
-  # If this point is reached, no RDS file exists so it has to be built from scratch
-  model <- load_ss_files(model_dir)
-  mcmc_dir <- file.path(model_dir, "mcmc")
-  if(dir.exists(mcmc_dir)){
-    mcmc_out <- SSgetMCMC(dir = mcmc_dir, writecsv = FALSE, verbose = TRUE)
-    model$mcmccalcs <- calc_mcmc(mcmc_out, ...)
+  if(!file.exists(rds_file)){
+    cat(green("Creating a new RDS file in", model_dir, "\n"))
+    # If this point is reached, no RDS file exists so it has to be built from scratch
+    model <- load_ss_files(model_dir)
+    model$mcmc_dir <- file.path(model_dir, "mcmc")
+    if(dir.exists(model$mcmc_dir)){
+      cat(green("Loading MCMC posteriors\n"))
+      mcmc_out <- SSgetMCMC(dir = model$mcmc_dir,
+                            writecsv = FALSE,
+                            verbose = FALSE)
+      model$mcmccalcs <- calc_mcmc(mcmc_out, ...)
+      cat(green(symbol$tick),
+          green(" Finished loading MCMC posteriors\n"))
+    }else{
+      cat(red(symbol$cross),
+          red(" MCMC posteriors not loaded (mcmc directory not found)\n"))
+    }
+    model$extra_mcmc_dir <- file.path(model_dir, "extra-mcmc")
+    if(dir.exists(model$extra_mcmc_dir)){
+      cat(green("Loading extra MCMC outputs\n"))
+      model$extra_mcmc <- fetch_extra_mcmc(model, ...)
+      cat(green(symbol$tick),
+          green(" Finished loading extra MCMC output\n"))
+    }else{
+      cat(red(symbol$cross),
+          red(" Extra MCMC outputs not loaded (extra-mcmc directory not found)\n"))
+    }
+    saveRDS(model, file = rds_file)
+    cat(green(symbol$tick),
+        green("Created RDS file successfully\n"))
   }
-
-  saveRDS(model, file = rds_file)
-  invisible()
+  rds_file
 }
 
 #' Load models from files created using [create_rds_file()]
@@ -50,11 +72,8 @@ load_ss_model_from_rds <- function(model_dir = NULL,
 
   verify_argument(model_dir, "character", 1)
 
-  model_rds_file <- file.path(model_dir, paste0(basename(model_dir), ".rds"))
-  if(!file.exists(model_rds_file)){
-    create_rds_file(model_dir, ...)
-  }
-  readRDS(model_rds_file)
+  rds_file <- create_rds_file(model_dir, ...)
+  readRDS(rds_file)
 }
 
 #' Load all the SS files for output and input, and return the model object
@@ -173,14 +192,15 @@ calc_mcmc <- function(mcmc,
   lst$sinit <- quantile(ssb %>% select(Initial) %>% pull(),
                         ss_mcmc_quants)
 
-  # sinit_post exists so that depletion calculations can be done for each posterior
-  sinit_post <- ssb %>% select(Initial)
+  # ssb_initial is added back later so that depletion calculations can be done
+  ssb_initial <- ssb %>% select(Initial) %>% pull()
 
-  ssb <- ssb %>% select(-c("Virgin", "unfished", "Btgt", "SPR", "MSY", "B_MSY/unfished"))
+  ssb <- ssb %>% select(-c("Virgin", "Initial", "unfished", "Btgt", "SPR", "MSY", "B_MSY/unfished"))
   lst$slower <- apply(ssb, 2, quantile, prob = lower)
   lst$smed   <- apply(ssb, 2, quantile, prob = med)
   lst$supper <- apply(ssb, 2, quantile, prob = upper)
 
+  ssb <- ssb %>% mutate(Initial = ssb_initial)
   depl <- ssb %>%
     mutate_at(.vars = vars(-Initial), .funs = ~{. / Initial}) %>%
     select(-Initial)
@@ -279,25 +299,64 @@ calc_mcmc <- function(mcmc,
 #' Load the SS model input and output data needed by this package in correct format
 #'
 #' @param ss_model SS3 model output as created by [create_rds_file()]
+#' @param s_yr Start year
+#' @param m_yr End year
+#' @param weight_factor A factor to multiply and divide SSB values by
 #' @param ... Arguments to be passed to [load_ss_sel_parameters()]
 #'
 #' @return A [list] with objects to be used throughout the package code
 load_ss_model_data <- function(ss_model,
+                               s_yr = NULL,
+                               m_yr = NULL,
+                               weight_factor = 1000,
                                ...){
 
   verify_argument(ss_model, "list")
+  verify_argument(s_yr, "numeric", 1)
+  verify_argument(m_yr, "numeric", 1)
+
   lst <- NULL
+  yrs <- s_yr:m_yr
 
   lst$parms_scalar <- load_ss_parameters(ss_model)
   lst$parms_sel <- load_ss_sel_parameters(ss_model,
                                           ...)
   lst$age_survey_df <- extract_age_comps(ss_model,
                                          age_comps_fleet = 2,
+                                         s_yr = s_yr,
+                                         m_yr = m_yr,
                                          ...)
   lst$age_catch_df <- extract_age_comps(ss_model,
                                         age_comps_fleet = 1,
+                                        s_yr = s_yr,
+                                        m_yr = m_yr,
                                         ...)
-  browser()
+
+  # The following is shown in a table in the assessment doc and made by the
+  # make.median.posterior.table() function in the hake-assessment repository
+  mc <- ss_model$mcmccalcs
+  extra_mc <- ss_model$extra_mcmc$timeseries
+  vals_mc <- mc[c("smed", "dmed", "rmed", "pmed", "fmed")]
+  vals_mc <- map(vals_mc, ~{
+    .x[names(.x) %in% yrs]
+  })
+  b <- extra_mc$Bio_all[extra_mc$Yr %in% yrs] / weight_factor
+  lst$med_popests <-  tibble(f_ssb = vals_mc$smed * weight_factor,
+                             rel_ssb = vals_mc$dmed,
+                             b = b,
+                             r = vals_mc$rmed * weight_factor,
+                             spr_f = vals_mc$pmed,
+                             e = vals_mc$fmed) %>%
+    mutate(yr = yrs) %>%
+    select(yr, everything())
+  # Make Relative fishing intensity and Exploitation fraction `NA` for the last year
+  # because the catch for that year has not occurred yet and the values output
+  # by the model are meaningless\
+  nrow_p <- nrow(lst$med_popests)
+  ncol_p <- ncol(lst$med_popests)
+  lst$med_popests[nrow_p, ncol_p - 1] <- NA
+  lst$med_popests[nrow_p, ncol_p] <- NA
+
   lst
 }
 
@@ -508,4 +567,328 @@ load_ss_sel_parameters <- function(ss_model = NULL,
     mutate(age = survey_ages)
 
   bind_rows(fish, survey)
+}
+
+#' Create and return a list of stats to attach to the main model by
+#' looking in the model's path for the report files.
+#'
+#' @param model The model object as output by [load_ss_files()]
+#' @param ss_mcmc_quants Quantile probability values to use
+#' @param ... Absorb arguments meant for other functions
+#' @return
+#' @importFrom readr read_table2 cols
+#' @importFrom purrr flatten
+#' @importFrom tibble add_column
+#' @importFrom dplyr group_nest slice
+#' @importFrom crayon yellow
+#' @export
+fetch_extra_mcmc <- function(model = NULL,
+                             ss_mcmc_quants = NULL,
+                             ...){
+
+  verify_argument(model, "list")
+  verify_argument(ss_mcmc_quants, "numeric", 3)
+
+  posts_file_name <- "posteriors.sso"
+
+  extra_mcmc_reports_dir <- file.path(model$extra_mcmc_dir, "reports")
+  extra_mcmc <- NULL
+
+  if(!dir.exists(model$extra_mcmc_dir)){
+    cat(red(symbol$cross),
+            red(" The", model$extra_mcmc_dir,
+                "directory does not exist, so the extra-mcmc wass not loaded\n"))
+    return(NULL)
+  }
+  if(!dir.exists(extra_mcmc_reports_dir)){
+    cat(red(symbol$cross),
+            red("The", extra_mcmc_reports_dir,
+                "directory does not exist, so the extra-mcmc wass not loaded\n"))
+    return(NULL)
+  }
+
+  ## Get the extra-mcmc directories
+  extra_mcmc_dirs <- dir(model$extra_mcmc_dir)
+  extra_mcmc_dirs <- file.path(model$extra_mcmc_dir,
+                               extra_mcmc_dirs[grepl("extra-mcmc", extra_mcmc_dirs)])
+
+  ## Get the number of Report.sso files in the directory
+  dir_list <- dir(extra_mcmc_reports_dir)
+  if(!length(dir_list)){
+    cat(red(symbol$cross),
+            red("There are no report files in the",
+                extra_mcmc_reports_dir, "directory\n"))
+    return(NULL)
+  }
+  report_files <- grep("^Report_[[:digit:]]+\\.sso$", dir_list)
+  num_reports <- length(report_files)
+  comp_files <- grep("^CompReport_[[:digit:]]+\\.sso$", dir_list)
+  num_comp_reports <- length(comp_files)
+  cat(green("Loading extra MCMC report files\n"))
+
+  ## Suppress warnings because there is an extra whitespace at the end of the header line in the file.
+  suppressWarnings(
+    posts <- read_table2(file.path(model$mcmc_dir, posts_file_name),
+                         col_types = cols())
+  )
+  ## Remove extra MLE run outputs. SS appends a new header followed by a 0-Iter row for an MLE run.
+  ## Sometimes MLEs are run by accident or on purpose at another time and forgotten about.
+  posts <- posts %>% filter(Iter != "Iter",
+                            Iter != 0)
+
+  ## Break up the loading of report files into the number of posteriors in each extra-mcmc subdir
+  num_reports_each <- map_int(extra_mcmc_dirs, ~{
+    ## Suppress warnings because there may be extra 'Iter' lines followed by '0' lines
+    ## because the SS MLE just appends these to the posteriors.sso file
+    suppressWarnings(
+      posts <- read_table2(file.path(.x, posts_file_name),
+                           col_types = cols())
+    )
+    posts <- posts %>% filter(Iter != "Iter",
+                              Iter != 0)
+    nrow(posts)
+  })
+
+  ## from_to is a two-column dataframe with the indices from and to for each processor to load report files
+  from_to <- tibble(from = 1, to = num_reports_each[1])
+  for(i in 2:length(num_reports_each)){
+    nxt <- tibble(from = from_to[i - 1,]$to + 1, to = from_to[i - 1,]$to + num_reports_each[i])
+    from_to <- bind_rows(from_to, nxt)
+  }
+
+  ## Load all report files into a list, 1 element for each report file. Elements that are NA had no file found
+  reps <- map(1:nrow(from_to), ~{
+    inds <- as.numeric(from_to[.x, 1]):as.numeric(from_to[.x, 2])
+    map(inds, ~{
+      rep_file <- file.path(extra_mcmc_reports_dir, paste0("Report_", .x, ".sso"))
+      if(!file.exists(rep_file)){
+        return(NA)
+      }
+      readLines(rep_file)
+    })
+  }) %>%
+    flatten()
+
+  ## Load all compreport files into a list, 1 element for each report file. Elements that are NA had no file found
+  compreps <- map(1:nrow(from_to), ~{
+    inds <- as.numeric(from_to[.x, 1]):as.numeric(from_to[.x, 2])
+    map(inds, ~{
+      comprep_file <- file.path(extra_mcmc_reports_dir, paste0("CompReport_", .x, ".sso"))
+      if(!file.exists(comprep_file)){
+        return(NA)
+      }
+      readLines(comprep_file)
+    })
+  }) %>%
+    flatten()
+
+  cat(green(symbol$tick),
+      green("Finished loading extra MCMC report files\n"))
+
+  ## Make custom reps_ objects for each output. Only relevant portions of the report file will be passed to
+  ## the table-making map2() calls later (speeds up the map2() calls)
+  rep_example <- reps[[which(!is.na(reps))[1]]]
+  comprep_example <- compreps[[which(!is.na(compreps))[1]]]
+  ## Biomass
+  bio_header_ind <- grep("^TIME_SERIES", rep_example) + 1
+  bio_header_line <- rep_example[bio_header_ind]
+  bio_header <- str_split(bio_header_line, " +")[[1]]
+  bio_start_ind <- bio_header_ind + 1
+  bio_end_ind <- grep("^SPR_series", rep_example) - 2
+  reps_bio <- map(reps, ~{.x[bio_start_ind:bio_end_ind]})
+  # Likelihood
+  like_start_ind <- grep("^LIKELIHOOD", rep_example) + 1
+  like_end_ind <- like_start_ind + 17
+  reps_like <- map(reps, ~{.x[like_start_ind:like_end_ind]})
+  ## Selectivity
+  next_yr <- model$endyr + 1
+  sel_header_ind <- grep("Factor Fleet Yr Seas", rep_example)
+  sel_header_line <- rep_example[sel_header_ind]
+  sel_header <- str_split(sel_header_line, " +")[[1]]
+  sel_ind <- grep(paste0(next_yr, "_1Asel"), rep_example)
+  reps_sel <- map(reps, ~{.x[sel_ind]})
+  ## Selectivity * Weight
+  selwt_ind <- grep(paste0(next_yr, "_1_sel\\*wt"), rep_example)
+  reps_selwt <- map(reps, ~{.x[selwt_ind]})
+  ## Natage
+  natage_header_ind <- grep("NUMBERS_AT_AGE_Annual_2 With_fishery", rep_example) + 1
+  natage_header <- str_split(rep_example[natage_header_ind], " +")[[1]]
+  natage_start_ind <- natage_header_ind + 1
+  natage_end_ind <- grep("Z_AT_AGE_Annual_2 With_fishery", rep_example) - 2
+  reps_natage <- map(reps, ~{.x[natage_start_ind:natage_end_ind]})
+  ## Q
+  q_header_ind <- grep("^INDEX_2", rep_example) + 1
+  q_header <- str_split(rep_example[q_header_ind], " +")[[1]]
+  q_start_ind <- q_header_ind + 1
+  ncpue <- nrow(model$dat$CPUE)
+  q_end_ind <- q_start_ind + ncpue - 1
+  reps_q <- map(reps, ~{.x[q_start_ind:q_end_ind]})
+  ## Comp tables
+  comp_header_ind <- grep("Composition_Database", comprep_example) + 1
+  comp_header <- str_split(comprep_example[comp_header_ind], " +")[[1]]
+  comp_start_ind <- comp_header_ind + 1
+  comp_end_ind <- grep("End_comp_data", comprep_example) - 1
+  reps_comp <- map(compreps, ~{.x[comp_start_ind:comp_end_ind]})
+
+  #' Extract the vectors from a list into a [tibble::tibble()]
+  #'
+  #' @param reps_lst A list of vectors, all the same length and structure,
+  #' typically extracted as a portion of a Report.sso file
+  #' @param header A vector of column names for the new table
+  #'
+  #' @return A [tibble::tibble()] representing one row for each of the list
+  #'  elements found in `reps_lst`. A new column called `Iter` is prepended and
+  #'  represents the list element number that the data for each row came from.
+  #'  List elements that are NA will not be included in the table.
+  extract_rep_table <- function(reps_lst, header){
+    lst <- map2(reps_lst, 1:length(reps_lst), ~{
+      if(is.na(.x[1])){
+        return(NULL)
+      }
+      vecs <- str_split(.x, " +")
+      vec_lengths <- map_int(vecs, ~{length(.x)})
+      vec_maxlength <- max(vec_lengths)
+      vecs <- map(vecs, ~{
+        length(.x) <- vec_maxlength
+        .x
+      })
+      tab <- do.call(rbind, vecs) %>%
+        as_tibble()
+      names(tab) <- header
+      tab %>%
+        add_column(Iter = .y, .before = 1)
+    })
+    do.call(rbind, lst) %>%
+      as_tibble()
+  }
+
+  sel <- extract_rep_table(reps_sel, sel_header) %>%
+    select(-c(2, 3, 5, 6, 7, 8)) %>%
+    map_df(as.numeric) %>%
+    filter(Yr == next_yr) %>%
+    select(-c(Iter, Yr))
+  selwt <- extract_rep_table(reps_selwt, sel_header) %>%
+    select(-c(2, 3, 5, 6, 7, 8)) %>%
+    map_df(as.numeric) %>%
+    filter(Yr == next_yr) %>%
+    select(-c(Iter, Yr))
+  natage <- extract_rep_table(reps_natage, natage_header) %>%
+    select(-c(2, 3)) %>%
+    map_df(as.numeric) %>%
+    filter(Yr == next_yr) %>%
+    select(-c(Iter, Yr))
+
+  ## Apply selectivity to numbers-at-age
+  natsel <- natage * sel
+  natselwt <- natage * selwt
+  extra_mcmc$natsel.prop <- natsel %>%
+    mutate(rsum = rowSums(.)) %>%
+    mutate_at(.vars = vars(-rsum), .funs = ~{.x / rsum}) %>%
+    select(-rsum)
+  extra_mcmc$natselwt.prop <- natselwt %>%
+    mutate(rsum = rowSums(.)) %>%
+    mutate_at(.vars = vars(-rsum), .funs = ~{.x / rsum}) %>%
+    select(-rsum)
+
+  # CPUE table and values (Q)
+  q <- extract_rep_table(reps_q, q_header) %>%
+    select(Iter, Exp, Calc_Q)
+  iter <- unique(q$Iter)
+  cpue <- q %>%
+    select(-Calc_Q) %>%
+    group_by(Iter) %>%
+    group_nest()
+  cpue <- do.call(cbind, cpue$data)
+  names(cpue) <- iter
+  extra_mcmc$cpue.table <- cpue %>%
+    as_tibble() %>%
+    map_df(~{as.numeric(.x)})
+
+  extra_mcmc$Q_vector <- q %>%
+    group_by(Iter) %>%
+    slice(1) %>%
+    pull(Calc_Q) %>%
+    as.numeric()
+
+  cpue <- apply(extra_mcmc$cpue.table,
+                MARGIN = 1,
+                FUN = function(x){quantile(as.numeric(x),
+                                           probs = ss_mcmc_quants)
+                })
+  extra_mcmc$cpue.0.025 <- as.numeric(cpue[1,])
+  extra_mcmc$cpue.median <- as.numeric(cpue[2,])
+  extra_mcmc$cpue.0.975 <- as.numeric(cpue[3,])
+
+  ## Add info on distribution of total biomass to existing time series data frame
+  timeseries <- extract_rep_table(reps_bio, bio_header) %>%
+    select(Iter, Bio_all, Bio_smry)
+  iter <- unique(timeseries$Iter)
+  Bio_all <- timeseries %>%
+    select(Iter, Bio_all) %>%
+    group_by(Iter) %>%
+    group_nest()
+  Bio_all <- do.call(cbind, Bio_all$data)
+  names(Bio_all) <- iter
+  Bio_all <- apply(Bio_all,
+                   MARGIN = 1,
+                   FUN = function(x){quantile(as.numeric(x),
+                                              probs = ss_mcmc_quants)
+                   })
+  extra_mcmc$timeseries <- model$timeseries
+  extra_mcmc$timeseries$Bio_all.0.025 <- as.numeric(Bio_all[1,])
+  extra_mcmc$timeseries$Bio_all.median <- as.numeric(Bio_all[2,])
+  extra_mcmc$timeseries$Bio_all.0.975 <- as.numeric(Bio_all[3,])
+
+  Bio_smry <- timeseries %>%
+    select(Iter, Bio_smry) %>%
+    group_by(Iter) %>%
+    group_nest()
+  Bio_smry <- do.call(cbind, Bio_smry$data)
+  names(Bio_smry) <- iter
+  Bio_smry <- apply(Bio_smry,
+                    MARGIN = 1,
+                    FUN = function(x){quantile(as.numeric(x),
+                                               probs = ss_mcmc_quants)
+                    })
+  extra_mcmc$timeseries$Bio_smry.0.025 <- as.numeric(Bio_smry[1,])
+  extra_mcmc$timeseries$Bio_smry.median <- as.numeric(Bio_smry[2,])
+  extra_mcmc$timeseries$Bio_smry.0.975 <- as.numeric(Bio_smry[3,])
+
+  comp <- extract_rep_table(reps_comp, comp_header)
+  ## median and quantiles of expected values and Pearsons
+  iter <- unique(comp$Iter)
+  comp <- comp %>%
+    filter(!is.na(N), N > 0)
+  exp_table <- comp %>%
+    select(Iter, Exp) %>%
+    group_by(Iter) %>%
+    group_nest()
+  exp_table <- do.call(cbind, exp_table$data)
+  names(exp_table) <- iter
+  exp_table <- apply(exp_table,
+                     MARGIN = 1,
+                     FUN = function(x){quantile(as.numeric(x),
+                                                probs = ss_mcmc_quants)
+                     })
+  extra_mcmc$agedbase <- model$agedbase
+  extra_mcmc$agedbase$Exp.025 <- exp_table[1,]
+  extra_mcmc$agedbase$Exp <- exp_table[2,]
+  extra_mcmc$agedbase$Exp.975 <- exp_table[3,]
+
+  pearson_table <- comp %>%
+    select(Iter, Pearson) %>%
+    group_by(Iter) %>%
+    group_nest()
+  pearson_table <- do.call(cbind, pearson_table$data)
+  names(pearson_table) <- iter
+  pearson_table <- apply(pearson_table,
+                         MARGIN = 1,
+                         FUN = function(x){quantile(as.numeric(x),
+                                                    probs = ss_mcmc_quants)
+                         })
+  extra_mcmc$agedbase$Pearson.025 <- pearson_table[1,]
+  extra_mcmc$agedbase$Pearson <- pearson_table[2,]
+  extra_mcmc$agedbase$Pearson.975 <- pearson_table[3,]
+
+  extra_mcmc
 }
